Detect redirect urls and update translations

+++ ui/src/main/java/io/snabble/sdk/ui/payment/Payone.kt
@@ -30,7 +30,7 @@ object Payone {
 
     @Parcelize
     data class PreAuthInfo(
-        val amount: Double?,
+        val amount: Int?,
         val currency: String?
     ) : Parcelable
